Refactor Team detail and fix design of card components.
