+++ frontend/src/Routes.js
@@ -19,7 +19,7 @@ import { TeamsAdminList} from "./pages/administration/Team/TeamsAdminList";
 import {ResetPassword} from "./pages/auth/ResetPassword";
 import {ResendToken} from "./pages/auth/ResendToken";
 import { AdministrationMenu } from './pages/administration/Administration/AdministrationMenu';
-import { User } from './pages/User';
+import { UserDetail } from './pages/users/detail/UserDetail';
 
 export function Routes() {
 	return (
@@ -33,7 +33,7 @@ export function Routes() {
             <Route path="/administration/teams" exact component={TeamsAdminList} />
 			<Route path="/teams/:id_team" exact component={TeamDetail} />
             <Route path="/teams" exact component={TeamList} />
-            <Route path="/user/:id_user" exact component={User} />
+            <Route path="/user/:id_user" exact component={UserDetail} />
 			<Route path="/leagues" exact component={Leagues}/>
 			<Route path="/matches" exact component={Matches}/>
 			<Route path="/statistics" exact component={Statistics} />
