+++ frontend/src/assets/sass/layout/_cards.scss
@@ -1,0 +1,9 @@
+.cardTitle {
+  font-size: 20px;
+  font-family: "Futura PT", sans-serif;
+}
+
+.card:hover {
+  box-shadow: 0 2px 22px 0 #eed600;
+  cursor: pointer;
+}
