+++ frontend/src/assets/sass/main.scss
@@ -5,6 +5,7 @@
 @import "layout/breadcrumb";
 @import "layout/advantageItems";
 @import "layout/reactTable";
+@import "layout/cards";
 
 @import "pages/userProfile";
 @import "pages/teamDetail";
