+++ frontend/src/templates/CardTemplate.js
@@ -1,10 +1,8 @@
 import {Card, CardDeck, Col, OverlayTrigger, Row, Tooltip} from 'react-bootstrap';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import React from 'react';
-import { useHistory } from "react-router";
-
-//temporary CSS
-const cardStyle = {cursor: "pointer"};
+import {useHistory} from "react-router";
+import {Heading} from "../atoms";
 
 export function CardTemplate({title, subtitle, pictureHeader, tooltipPictureHeader, mainPicture, textHeader, redirect}) {
     let history = useHistory();
@@ -13,18 +11,17 @@ export function CardTemplate({title, subtitle, pictureHeader, tooltipPictureHead
         history.push(redirect);
     }
 
-
     return (
-        <Col lg={4} md={6} sm={12} xs={12} className="mt-4">
+        <Col lg={4} md={6} sm={12} className="mt-4">
             <CardDeck>
-                <Card style={cardStyle} onClick={onRedirect}>
+                <Card onClick={onRedirect}>
                     <Card.Header>
                         <Row className="align-items-center">
-                            <Col lg={8} md={8} sm={8} xs={9}>
-                                <Row>{title}</Row>
+                            <Col lg={10} sm={10} xs={9}>
+                                <Row><Heading size="md" className="cardTitle">{title}</Heading></Row>
                                 <Row>{subtitle}</Row>
                             </Col>
-                            <Col className="text-right">
+                            <Col lg={2} sm={2} xs={3} className="text-right">
                                 <OverlayTrigger
                                     overlay={<Tooltip id="tooltip-disabled">{tooltipPictureHeader}</Tooltip>}>
                                     <FontAwesomeIcon icon={pictureHeader} size="2x"/>
